Move ToUnstructured function from clusterctl to util
